requirements.txt
Brotli==1.0.7
dj-database-url==0.5.0
Django==2.2.3
django-crispy-forms==1.7.2
django-location-picker==0.2.0
django-phone-field==1.7.1
gunicorn==19.9.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.3
wrapt==1.11.1
